RETURN

+++ btconfig.h
@@ -7,7 +7,8 @@ extern size_t cfg_req_slice_size;
 extern size_t quit_after_download;
 
 #define MAX_METAINFO_FILESIZ	(4*1024*1024)
-#define DEFAULT_SLICE_SIZE (16*1024u)
+#define DEFAULT_SLICE_SIZE (16*1024)
+
 // According to specs the max slice size is 128K.  While most clients now do
 // not accept a value that large, we want max compatibility.
 #define cfg_max_slice_size (128*1024)
@@ -16,7 +17,7 @@ extern size_t cfg_req_queue_length;
 #define PEER_ID_LEN 20
 #define PEER_PFX "-CD0303-"
 
-extern char *cfg_user_agent; // HTTP header
+extern char *cfg_user_agent;	// HTTP header
 
 extern size_t cfg_cache_size;
 
@@ -33,14 +34,15 @@ extern time_t cfg_seed_hours;
 extern double cfg_seed_ratio;
 
 extern int cfg_max_bandwidth_down;
-extern ssize_t cfg_max_bandwidth_up;
+extern int cfg_max_bandwidth_up;
+
 extern size_t arg_hash_fails;
 
 // arguments global value
 extern char *arg_metainfo_file;
 extern char *arg_bitfield_file;
 extern char *arg_save_as;
-extern char *arg_user_agent; // BT peer ID, not HTTP header
+extern char *arg_user_agent;	// BT peer ID, not HTTP header
 
 extern unsigned char arg_flg_force_seed_mode;
 extern unsigned char arg_flg_check_only;
